backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
prisma==0.15.0
redis==5.2.1
celery[redis]==5.4.0
boto3==1.35.76
stripe==11.3.0
httpx==0.28.1
aiohttp==3.11.10
tenacity==9.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
alembic==1.14.0
sqlalchemy==2.0.36
prometheus-client==0.21.1
sentry-sdk[fastapi]==2.19.2
pillow==11.0.0
numpy==2.2.1
scipy==1.15.0
scikit-learn==1.6.1
pandas==2.2.3
matplotlib==3.10.0
seaborn==0.13.2
pytest==8.3.4
pytest-asyncio==0.25.2
black==24.10.0
ruff==0.9.1